IBoutlets as private var
