--- ImageFeed/Helpers + Extensions/ColorSet.swift
@@ -1,14 +1,0 @@
-//
-//  ColorSet.swift
-//  ImageFeed
-//
-//  Created by Федор Завьялов on 06.01.2024.
-//
-
-import UIKit
-
-extension UIColor {
-    static var ypblack: UIColor { UIColor(named: "YP Black") ?? UIColor.black }
-    static var ypwhite: UIColor { UIColor(named: "YP White") ?? UIColor.white }
-    static var ypred: UIColor { UIColor(named: "YP Red") ?? UIColor.red }
-}
