+++ ImageFeed/Helpers + Extensions/Colors.swift
@@ -7,11 +7,10 @@
 
 import UIKit
 
-extension UIColor{
-    
+extension UIColor {
     static var ypBlack: UIColor {UIColor(named: "YP Black") ?? UIColor.black}
     static var ypWhite: UIColor {UIColor(named: "YP White") ?? UIColor.black}
     static var ypRed: UIColor {UIColor(named: "YP Red") ?? UIColor.red}
-    static var transperantWhite: UIColor {UIColor(named: "Transperant White") ?? UIColor.white}
+    static var transperantWhite: UIColor {UIColor(named: "Transperant White") ?? UIColor.white }
 }
